Firmenstamm-Verwendungen an aktive Ansicht anbinden
