--- src/renderer/views/FirmsLegacyView.actual
@@ -1,1 +1,0 @@
-placeholder
