--- src/renderer/views/FirmsLegacyView2.js
@@ -1,1 +1,0 @@
-x
