--- src/renderer/views/FirmsView.new
@@ -1,1 +1,0 @@
-export { default } from "./FirmsUsageView.js";
